fix(capture): share one kafka error label vocabulary for warnings delivery
The delivery-errors counter labeled errors with the RDKafkaErrorCode variant name, allocating a String per failed delivery on rdkafka's poll thread and reporting MessageTimedOut where every other capture metric reports message_timed_out. Move capture's curated error_code_tag into common/kafka so both sides share it, and stamp path so the counter splits per pipeline like ingestion_warnings_total.
