+++ rust/common/kafka/Cargo.toml
@@ -21,3 +21,6 @@ tracing = { workspace = true }
 uuid = { workspace = true }
 thiserror = { workspace = true }
 tokio = { workspace = true }
+
+[dev-dependencies]
+rstest = "0.26"
