+++ rust/common/kafka/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod config;
+pub mod error;
 pub mod kafka_consumer;
 pub mod kafka_messages;
 pub mod kafka_producer;
